+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This file documents recent notable changes to this project. The format of this
 file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Made `giganto_name` required and removed the fallback that used the Giganto
+  address IP as the TLS server name.
+
 ## [0.9.0] - 2026-06-19
 
 ### Added
@@ -284,6 +291,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Send the generated timeseries to Giganto's ingest.
 - Save the model's id and the last time the timeseries was sent to a file.
 
+[Unreleased]: https://github.com/aicers/crusher/compare/0.9.0...main
 [0.9.0]: https://github.com/aicers/crusher/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/aicers/crusher/compare/0.7.1...0.8.0
 [0.7.1]: https://github.com/aicers/crusher/compare/0.7.0...0.7.1